移動・補間、アニメーション等のカウント更新を時間に置き換え

+++ DXLibProject_Start/World/Character/Guardian/DragonAttack.h
@@ -14,6 +14,6 @@ public:
 private:
 	bool m_active;
 	bool m_attackFlag;
-	int m_attackCount;
+	float m_attackCount;
 };
